quan_ly_tien_do_web_v2/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.25
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
alembic>=1.13.1
python-multipart>=0.0.6
pandas>=2.3.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
websockets>=12.0
google-generativeai>=0.4.0
